Update patch set 1
Patch Set 1:

Thanks for the nudge about a newer RMR version, I'll get on it.

FYI, Alpine is a whole world unto itself.  It is NOT a Debian distribution, not a Centos distribution either.  Yes to your point, Alpine uses musl instead of libc.

When checking tag files please be sure you're in the right parent directory :/

Patch-set: 1