Update patch set 1
Patch Set 1:

my preference is that you make an orthogonal changeset for the image as that's not related to this change. 

The alpine Dockerfile and non alpine dockerfiles need not be similar since we do not have an APK storage but we do have deb/rpm etc. So, do what is best for the non alpine image.

Patch-set: 1