Update patch set 2
Patch Set 2:

Never released. The base images are pushed to the LF docker "staging" registry nexus3.o-ran-sc.org:10004 and refreshed weekly. The staging registry allows regular pushes (overwrites) so that the RIC base images keep up with changes in their respective base images. That is not possible in a release registry.

Patch-set: 2