Update patch set 2
Patch Set 2:

refresh: the merge job is triggered weekly by a cron timer. That rebuilds the image and pushes it to the Nexus3 staging registry, overwriting the old image.  This picks up updates in base images such as Alpine and Ubuntu. The set of software and especially the RMR version does not change, but depending on the base image, Python/Go/etc. could change.

Patch-set: 2