Update patch set 1
Patch Set 1:

There is a change coming to RMR in the next day or so (requires a bit of stress testing before I will bless it) that eliminates the max message size for received messages.  This is needed to support things like E2-Term which receive (astonishingly, insanely) huge messages from some devices.   I don't mind giving a +1 here, but it will be important to pick this change up once it is released, so it might save some effort to wait.   The upcoming change should be 3.7.1.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>