Update patch set 1
Patch Set 1:

(5 comments)

Some comments in constants with regard to schema.  A comment there too, but applies globally to this change:  I notice only AT&T in the copyright; is this not considered joint work with Nokia, or does it fall into the "it's an xAPP" category?

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>