Update patch set 1
Patch Set 1:

RMR 4.0.0 is close.  The change is in gerrit and I'm working on getting the unit tests to run via jenkins (something is still referencing NNG when it shouldn't be I think).   Probably makes sense just to wait for that as it will eliminate all of the NNG things.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>