Update patch set 1
Patch Set 1:

Scott, let's use this as the test case for 4.0.0.  Merge this, I'll update A1 to use it, and we'll go from there.  The docker image is about 125KB so it's not like we're emptying the bitbucket for this :)

Patch-set: 1
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0